make only video owner can edit,delete

+++ src/controllers/videoController.js
@@ -56,18 +56,21 @@ export const postUpload = async (req, res) => {
 export const watch = async (req, res) => {
     const { id } = req.params;
     const video = await Video.findById(id).populate("owner");
-    console.log(video);
     if (!video) {
         return res.status(404).render('404', { pageTitle: 'Video Not Found' });
     }
-    return res.render('watch', { pageTitle: video.title, video, owner });
+    return res.render('watch', { pageTitle: video.title, video });
 }
 
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id } } = req.session;
     const video = await Video.findById(id);
     if (!video) {
         return res.status(404).render('404', { pageTitle: 'Video Not Found' });
+    }
+    if (String(video.owner) !== _id) {
+        return res.status(403).redirect('/');
     }
     return res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
 }
@@ -78,6 +81,9 @@ export const postEdit = async (req, res) => {
     const video = await Video.exists({ _id: id });
     if (!video) {
         return res.status(404).render('404', { pageTitle: 'Video Not Found' });
+    }
+    if (String(video.owner) !== _id) {
+        return res.status(403).redirect('/');
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -90,6 +96,14 @@ export const postEdit = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id } } = req.session;
+    const video = await Video.findById(id);
+    if (!video) {
+        return res.status(404).render('404', { pageTitle: 'Video Not Found' });
+    }
+    if (String(video.owner) !== _id) {
+        return res.status(403).redirect('/');
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect('/');
 }
